hide ascii logo on mobile, clean up cli demos
Amp-Thread-ID: https://ampcode.com/threads/T-019c407b-3bf1-7067-836a-65a118ce757e
